pyproject.toml
[tool.poetry]
name = "aule_unibo"
version = "0.1.0"
description = ""
authors = ["VaiTon <eyadlorenzo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.12"
requests = "^2.27.1"
python-dateutil = "^2.8.2"
tzlocal = "^4.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
